+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.2a1]
 ### Added
 - Option to have the ComponentClient be event managed when linked to an event manager.
   This is True by default.
@@ -33,5 +35,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Iffy behaviour around "locking" the reaction executor which lead to some requests just being ignored.
 
-[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.0.1a1...HEAD
+[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.0.2a1...HEAD
+[1.0.2a1]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.0.1a1...v1.0.2a1
 [1.0.1a1]: https://github.com/olivierlacan/keep-a-changelog/compare/0.0.2...v1.0.1a1